move routing job tests to separate file
