+++ validator/app/src/compute_horde_validator/validator/tests/test_facilitator_client.py
@@ -9,19 +9,15 @@ from unittest.mock import patch
 import pytest
 import websockets
 from channels.layers import get_channel_layer
-from compute_horde.executor_class import DEFAULT_EXECUTOR_CLASS, ExecutorClass
+from compute_horde.executor_class import DEFAULT_EXECUTOR_CLASS
 from compute_horde.fv_protocol.facilitator_requests import (
     JobRequest,
     Response,
-    V2JobRequest,
 )
 from compute_horde.fv_protocol.validator_requests import (
     V0AuthenticationRequest,
     V0MachineSpecsUpdate,
 )
-from compute_horde.receipts.models import JobStartedReceipt
-from compute_horde.receipts.schemas import JobStartedReceiptPayload, Receipt
-from compute_horde.utils import sign_blob
 from django.utils import timezone
 
 from compute_horde_validator.validator.models import (
@@ -36,8 +32,6 @@ from compute_horde_validator.validator.organic_jobs.facilitator_client import (
 from compute_horde_validator.validator.organic_jobs.miner_driver import JobStatusUpdate
 from compute_horde_validator.validator.utils import MACHINE_SPEC_CHANNEL
 
-from ..organic_jobs import routing
-from ..organic_jobs.routing import pick_miner_for_job
 from .helpers import (
     MockFaillingMinerClient,
     MockMetagraph,
@@ -312,81 +306,6 @@ class FacilitatorExpectMachineSpecsWs(FacilitatorWs):
             else:
                 async with self.condition:
                     self.condition.notify()
-
-
-@pytest.mark.django_db(databases=["default", "default_alias"], transaction=True)
-@pytest.mark.asyncio
-async def test_pick_miner_for_job__no_matching_executor_class():
-    await setup_db()
-    with pytest.raises(routing.NoMinerForExecutorType):
-        await pick_miner_for_job(
-            V2JobRequest(
-                uuid=str(uuid.uuid4()),
-                executor_class=next(c for c in ExecutorClass if c != DEFAULT_EXECUTOR_CLASS),
-                docker_image="doesntmatter",
-                raw_script="doesntmatter",
-                args=[],
-                env={},
-                use_gpu=False,
-            )
-        )
-
-
-@pytest.mark.django_db(databases=["default", "default_alias"], transaction=True)
-@pytest.mark.asyncio
-async def test_pick_miner_for_job__no_online_executors():
-    await setup_db()
-    await MinerManifest.objects.all().aupdate(online_executor_count=0)
-    with pytest.raises(routing.NoMinerForExecutorType):
-        await pick_miner_for_job(
-            V2JobRequest(
-                uuid=str(uuid.uuid4()),
-                executor_class=DEFAULT_EXECUTOR_CLASS,
-                docker_image="doesntmatter",
-                raw_script="doesntmatter",
-                args=[],
-                env={},
-                use_gpu=False,
-            )
-        )
-
-
-@pytest.mark.django_db(databases=["default", "default_alias"], transaction=True)
-@pytest.mark.asyncio
-async def test_pick_miner_for_job__all_executors_busy(validator_keypair, miner_keypair):
-    await setup_db()
-    async for miner in Miner.objects.all():
-        for _ in range(5):
-            payload = JobStartedReceiptPayload(
-                job_uuid=str(uuid.uuid4()),
-                miner_hotkey=miner.hotkey,
-                validator_hotkey=validator_keypair.ss58_address,
-                timestamp=timezone.now(),
-                executor_class=DEFAULT_EXECUTOR_CLASS,
-                max_timeout=60,
-                is_organic=True,
-                ttl=60,
-            )
-            blob = payload.blob_for_signing()
-            receipt = Receipt(
-                payload=payload,
-                validator_signature=sign_blob(validator_keypair, blob),
-                miner_signature=sign_blob(miner_keypair, blob),
-            )
-            await JobStartedReceipt.from_receipt(receipt).asave()
-
-        with pytest.raises(routing.AllMinersBusy):
-            await pick_miner_for_job(
-                V2JobRequest(
-                    uuid=str(uuid.uuid4()),
-                    executor_class=DEFAULT_EXECUTOR_CLASS,
-                    docker_image="doesntmatter",
-                    raw_script="doesntmatter",
-                    args=[],
-                    env={},
-                    use_gpu=False,
-                )
-            )
 
 
 # TODO: this test is flaky, needs proper investigation
